Refactor pet model to change types for age, weight as string

+++ internal/handler/pet_handler.go
@@ -57,24 +57,26 @@ func (h *PetHandler) Create(c *fiber.Ctx) error {
 	form, err := c.MultipartForm()
 	if err == nil {
 		files := form.File["images"]
-		for _, file := range files {
-			ext := filepath.Ext(file.Filename)
-			newFilename := fmt.Sprintf("%d-%s%s", time.Now().Unix(), util.GenerateRandomHash(12), ext)
-			filePath := filepath.Join(uploadDir, newFilename)
+		if len(files) > 0 {
+			for _, file := range files {
+				ext := filepath.Ext(file.Filename)
+				newFilename := fmt.Sprintf("%d-%s%s", time.Now().Unix(), util.GenerateRandomHash(12), ext)
+				filePath := filepath.Join(uploadDir, newFilename)
 
-			if err := c.SaveFile(file, filePath); err != nil {
-				// Remove if not save
-				if len(savedFiles) > 0 {
-					for _, savedFile := range savedFiles {
-						os.Remove(savedFile)
+				if err := c.SaveFile(file, filePath); err != nil {
+					// Remove if not save
+					if len(savedFiles) > 0 {
+						for _, savedFile := range savedFiles {
+							os.Remove(savedFile)
+						}
 					}
+
+					return c.Status(fiber.StatusInternalServerError).JSON(Response{Error: true, Message: "Could not save file"})
 				}
 
-				return c.Status(fiber.StatusInternalServerError).JSON(Response{Error: true, Message: "Could not save file"})
+				imagesPath = append(imagesPath, "/"+filePath)
+				savedFiles = append(savedFiles, filePath)
 			}
-
-			imagesPath = append(imagesPath, "/"+filePath)
-			savedFiles = append(savedFiles, filePath)
 		}
 	}
 
@@ -277,10 +279,10 @@ func (h *PetHandler) Update(c *fiber.Ctx) error {
 	if body.Name != "" {
 		updateFields["name"] = body.Name
 	}
-	if body.Age > 0 {
+	if body.Age != "" {
 		updateFields["Age"] = body.Age
 	}
-	if body.Weight > 0 {
+	if body.Weight != "" {
 		updateFields["Weight"] = body.Weight
 	}
 	if body.Color != "" {
